Fixed plan initialization missing
We forgot to call `plan.Init()` after `makePlan(...)` which caused the scanners to not be properly initialized as there was no propagation. This was incorrectly "fixed" without realizing it by checking if the scanners had beed initialized on each `scan.Next()` call.
This worked, but added constant overhead to each call, and could be resolved by properly initializing the entire plan graph once at the beginning.