refactor: Rework client.DB to ensure interface contains only public types (#277)
* Rename var to collection name

Is clearer

* Rename GetCollection=>GetCollectionByName

Previous fn name was a bit ambigious, particularly due to the existance of GetCollectionBySchemaID, which has the same method signature.

* Remove schemaManager from DB interface

Exposes a large amount of internal generator types and is not nice on a public interface

* Use internal multistore instead of public fn

Pub fn is an oddity that might be removed, this is one of two refs to it

* Use blockstore over dagstore for public interfaces

Removes the defra-only DagStore type/name from the db interface - blockstore concept appears to be larger than defra, and there seems no sensible way to remove it from the interfaces (p2p lib requires one).  If Cid and Block should be considered standard, then blockstore should be too.

* Remove printDebugDb

Is a duplicate of the public PrintDump

* Remove public XYZStore accessors

None of these need to be public.  Leaves us with the (standard-typed, standard-concept) Root and Block stores only.

* Move store and txn interfaces into client

txn is dependent on store, and txn is on the public db+collection interfaces

* Reorder the DB interface

* Remove unused sequence interface

* Remove schema from document

Is not used, and creates extra dependency headaches when moving public items into client package.

* Remove blockstore.GetBlock

* Rename datastores to datastore

* Move store package to datastore dir

* Move store.go into datastore package

* Remove unwanted comment