tools: Add github PR template. (#394)
- RELEVANT ISSUE(S)

Resolves: https://github.com/sourcenetwork/defradb/issues/393

- DESCRIPTION

Adds the following PR template:

```
## RELEVANT ISSUE(S)

Resolves #

## DESCRIPTION

Please include a summary of the change and which issue is fixed. If there isn't an issue, then please create one.
Please also include relevant motivation and context. List any dependencies that are required for this change.

### HOW HAS THIS BEEN TESTED?

Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.

### CHECKLIST:

- [ ] I have commented the code, particularly in hard-to-understand areas.
- [ ] I have made corresponding changes to the repo-held documentation.
- [ ] I have made sure that the PR title adheres to the conventional commit style (subset of the ones we use can be found under: [tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)

### ENVIRONMENT / OS THIS WAS TESTED ON?

Please specify which of the following was this tested on (remove or add your own):
- [ ] Arch Linux
- [ ] Debian Linux
- [ ] MacOS
- [ ] Windows
```