refactor: Internal doc restructure (#471)
* Alias internal doc object

* Reduce Select mutation

* Remove unused return param (arb. join)

* Remove extra funcs from MultiNode

* Remove deprecated unit tests

Are hard to maintain with the coming changes, and prod code should be covered by integration tests

* Remove connor unit tests

Maintaining them is not worth the effort with connor now being covered by our integration tests, and the upcoming changes.

* Restructure internal document data structure

* Rework type join to make use of new select

Work could be taken futher by creating a new mapper.relation struct, but would suggest that is out of scope here. Also removes an explain parameter as it is not useful, duplicating info already highly visible in the result whilst using an otherwise dead property.

* Remove dead code from parser package

* I443: Expand deep group filter tests

Looks like case A in the issue was already tested, and the new test passes in develop - so I guess this was just a coverage issue, unless something else fixed it before I448.

* I387: Handle multiple aggregate sources

* I390: Handle multiple aliased joins with filters

Checked test against develop, it failed.  Was a bug and it is now fixed.

* I484: Handle filters on unrendered children

* I495: Handle version select within create

Turned out the bug wasnt really related to this PR (unless there was another bug hiding this one maybe), but I fixed it anyway.