refactor: Remove some possible panics from codebase (#732)
* Assert span implements Span

* Remove dead span code

* Handle db.PrintDump errors

* Simplify doc container

numDocs property adds no value, and adds a new potential source of failure (if it becomes out of sync with the actual slice length).  Panics add no value as the panic that would otherwise happen (index out of bounds) is plenty clear enough anyway.  Returning error is not desired here as this struct is esentially an extension of a slice, and like slices index validation should be handled by callers - allowing them to determine whether it is safe or not to access elements without additional validation.

* Remove unused utils function