chore: Remove `ipfslite` dependency (#739)
Relevant issue(s)
Resolves #738

Description
This PR removes our direct dependency on ipfslite. It is still an indirect dependency simply because it's used by an other package we import. Not from a function that we use but from a test of that package.

Note that some commented code will need to be uncommented once #683 is resolved. Although a ds.Batching implementing datastore needs to be used for the DHT within the net package, as long as badgerds.Datastore keeps the Batch method, the should have no problem.