refactor: Cleanup parsing logic (#909)
* Remove unused QueryType prop

* Remove unused order.statement prop

* Replace optionalDocKeys with client.Option

Type was added before we had client.Option

* Remove legacy commit entry

Should have been removed in the PR that removed the commit query

* Remove unused statement property

* Remove unused func from interface

* Remove Root from Field

* Remove unused func

* Parse aggregates in parser package

* Remove statement from parser.Select

* Remove ast from OperationDefinition

* Remove unused Name from OperationDefinition

* Make parser.Alias optional

The parser models will shortly be exposed as part of the public interface, and it should not be on the users to have to always populate this property.  It also makes more sense IMO to do it this way regardless, as alias is optional, but render-keys are not.

* Make commit.DocKey option type

Makes it far clearer that it is optional

* Use option for parser.limit

Also flattens them, as the previous structure suggested if one was present then the other was mandatory. Also switches from int64 to uint64 as they should not be negative (and makes it more consitent with a few other similar props who use uint64).

* Use option for parser.order

* Remove '.'s from parser.order.fields

No reason for this to exist, and will confuse anyone using it.

* Use option for parser.GroupBy

* Use option for parser.Filter

* Cleanup commit-query branching

Now using proper consts, and defined in the same location as everything else.

* Remove unused param

* Remove unused commit.GetRoot func

* Remove unused mutation.GetRoot func

* Tidy up parser.Root references

Previous if was conceptually incorrect. None option added no value (select.Root now defaults to Object, which could be handy for users anyway).  Also types the enum.

* Move parser.types to client dir

Also breaks up types.go in preparation for the moving of the rest of the model.

* Rename parser.Query to parser.Request

As agreed as a team, commit does not seek to rename all variables referencing this - that can be done later to save drowning this large PR with minor changes

* Move request model out of parser package

The parser package is now responisble for converting an gql-ast into a shared model.  Soon it will be responisble for converting a string into the shared model.  Mostly a copy-paste commit.