Extract some cli errors
I realized that continuing this in this package would be ineffiecient as Orpheus is reworking it at the moment.  He said (without seeing these changes) that it was still useful and shouldnt clash too hard, but I think it would be better to wait a bit to do the rest of the package.