Rework replicator tests so that each expected result is asserted upon
Deletion of pretty much any property/item in the test case will now cause it to fail, before many of these deletions would permit it to pass without asserting the expected results.

Much of this commit has been moved and reworked later, however I chose to preserve it as it highlights the behavioural change specific to this commit.