Refactor test library
This commit merges the properties associated with the peer-tests with the properties associated with the replicator tests.  There was a lot of duplication, and it prevented testing of both peer and replicator tests within the same test case due to id/index clashes.  It also reworks the code so that the two elements work in largely the same way. It should allow replicators to test updates, and peers to test creates.

The both peer and replicator tests now assert on the final state of all databases for all nodes. This feels both more meaningful and clearer IMO, although it does add a few lines to each test, as well as some fanciness required to deal with the non-deterministic results of the peer sync tests.