refactor: Allow db keys to handle multiple schema versions (#1026)
* Rename collectionKey variable

Is clearer

* Add schema version key

Name will likely be improved with https://github.com/sourcenetwork/defradb/issues/1025

* Remove duplicate variable

* Deduplicate business logic

These absolutely have to have the same value, the descriptive new variable name will also come in handy shortly

* Document global-local distinction of schema elements

* Move schema id evaluation to before saving of collectionKey

SchemaId will be used first shortly and I dont want to polute the logic-change commit with move stuff

* Rename collectionSchemaKey variable

Is clearer

* Persist schemaVersionKey

* Persist schemaVersionId at schemaKey

SchemaKey is now a pointer to the current version, essentially becoming schema.head. getCollectionByVersionId is a copy-paste of the existing GetCollectionByName function. GetCollectionByName will be reworked shortly to work in a similar fashion.

* Persist schemaVersionId at collectionKey

Uses the getCollectionByVersionId defined in an earlier commit. getCollectionDesc is an existing duplicate function that as well as duplicating itself, also duplicates much of the logic in getCollectionByVersionId - it should be refactored at somepoint, but correcting this is IMO out of scope at the moment.