fix: Rework the P2P integration tests (#989)
* Duplicate net tests

In order to preserve the order driven tests

* Rework replicator tests so that each expected result is asserted upon

Deletion of pretty much any property/item in the test case will now cause it to fail, before many of these deletions would permit it to pass without asserting the expected results.

Much of this commit has been moved and reworked later, however I chose to preserve it as it highlights the behavioural change specific to this commit.

* Expand replicator tests

There was a typo, and the name suggests it tests everything replicator related - and that no further tests are required

* Refactor test library

This commit merges the properties associated with the peer-tests with the properties associated with the replicator tests.  There was a lot of duplication, and it prevented testing of both peer and replicator tests within the same test case due to id/index clashes.  It also reworks the code so that the two elements work in largely the same way. It should allow replicators to test updates, and peers to test creates.

The both peer and replicator tests now assert on the final state of all databases for all nodes. This feels both more meaningful and clearer IMO, although it does add a few lines to each test, as well as some fanciness required to deal with the non-deterministic results of the peer sync tests.

* Remove uneeded check

Looks both overly defensive, and enables a silent failure

* Breakup P2P test file

Is about to get crowded :)

* Expand tests