refactor: Change terminology from query to request (#1054)
- Resolves #455

- DESCRIPTION:

This PR changes existing terminology to comply with the following terminology definitions:

DefraDB Terminology:
* (1) `Request` is the top-level term that encapsulates all the operations (`query` and `mutation` ops).
* (2) `Query Request` or `Request Query` are same as (1) but only for situations where we need to use the word "Query" for some reason (say for external use for example in `cli` package), so the `Request` descriptor clarifies the type of "Query" this is.
* (3) `Query Operation` implies that the request is a "read-only request", but note that if we loosely say `Query Request` that does not mean a read-only operation according to (2).
* (4) `Query` loosely the same as (3), BUT STRONGLY AVOID USING ON IT'S OWN.
* (5) `Mutation Operation` implies that the request is a "write request".
* (6) `Mutation Request` implies that the request's operation is of type `mutation`. Note: `(6)` and `(5)` can be used interchangeably but `(2)` and `(3)` can not.
* (7) `Mutation` is loosely the same as (5) and (6) but avoid using this term.
 
With the above definition, we can use other descriptors to specify the type of request it is, for example: "Explain request" or "Subscription Request". But the only gotcha / confusing rule we need to worry about is that `Query Request` is not the same as `Query Operation`. The use of the word `Query` with another descriptor assumes (4), for example: `Query Plan` is a plan that is made from a read-only query.  

Throughout the codebase (excluding 3rd party stuff, eg. ipfs), we should avoid the use of (2), (4), and (7) to avoid confusion. The only exception to (4) I can see is when we say "Defra Query Language" (because this is not referring to the query operation).