feat: Persist schema version at time of commit (#1055)
* Remove unused dag.GetSchemaID func

* Break up long lines

Will be added to shortly

* BREAKING - Remove unused SchemaId prop

Breaks some tests as although the property is always empty it is accounted for in CID generation.

* BREAKING - Remove unused Schema.Key prop

Note: The empty unused prop was used to construct some cids, removing it thus changes the cids even though it was always empty.

* Replace commit schemaId with versionKey

The primary driving change in this commit is the addition of SchemaVersionKey to CompositeDAGDelta in core/crdt/composite.go, this persists the schema state alongside the document change.  All other changes in this commit are to facilitate this change.

This commit got a bit more involved than I intended as getCollectionByVersionId was regenerating the schema id instead of taking it from the object it already had - possibly because the property on the object that it already had never actualy set the property.  This got a bit awkward as the id lives on the object it is generated from - when the prop was populated on regeneration of itself it was taken into account as part of the regenerated key causing it to not match the originally (saved) key.

Cids in the tests have been corrected in this commit for the production code changes in this commit and the prior BREAKING commits (to save fixing them 3 times).