Generate gql types using collection descriptions
Simplifies the flow somewhat and allows our strongly typed model to be the primary driver.  This allows more work to leave the gql space, including the persisted [gql]schema in the datastore allowing for greater decoupling of Defra from gql. It also frees the generator code from having to worry about any internal database stuff such as relationship management, focusing on just gql stuff.

It also means that any Json patches applied to the persisted json schema (in /collection/version) will automatically be applied to the query language (gql) types. This /collection/version key will require some changes when we start to introduce local-only elements, however that is not a new problem to this PR.

It also decouples the gql types from the gql sdl, in theory allowing a non-gql sdl to generate gql types, or a gql sdl to generate types for a non-gql query language.