Create new integration test framework
Creates a new integration test framework, and migrates shared functions to its files from the old framework. The Init code is unchanged, as is the change detector code - they have just moved.

Is not in use until the next commit - I thought it might be easier to review if split into two. It provides the same features as the old system, but without implicit/ridged order of operations, and hopefully simpler code.

Bonus fix - transactional tests now work with the change detector.