Make old test framework to use new framework
Changes the old test framework so that it is essentially syntax sugar on top of the new framework.

This is temporary, and is done to allow us to introduce (and review) the new framework without having to migrate all our existing tests to the new framework. The old framework should be removed as soon as is convienent, and can be done so bit by bit. It also ensures that the new framework contains all the functionality of the old framework.

Explain and collection tests use a different framework. Those frameworks have not been migrated here, but their tests should be migrated at some point - it should be fairly straight forward to extend the new test framework to do so.

New integration tests should be written using the new framework only. This includes collection tests and explain tests.