feat: Add P2P collection topic subscription (#1086)
Relevant issue(s)
Resolves #1081

Description
This PR adds the option to subscribe to events on a given collection by using the collectionID (in reality schemaID for now as there is no globalSchemaID or anything similar) as the pubsub topic. It adds commands to the CLI to add, remove and list the P2P collection topics.