feat: Add `like` sub-string filter (#1091)
Relevant issue(s)
Resolves #982

Description
This PR adds _like and _nlike filters to match on a sub-string basis. The % character can be used before and/or after the string to specify how the string should be matched.

"String to match" will try to match the string exactly. It is equivalent to using the _eq filter.
"%String to match" will match to a string ending with String to match.
"String to match%" will match to a string starting with String to match.
"%String to match%" will match to a string that contains String to match.