PR FIXUP - Assert expected error raised
Also converts one of the transaction tests, as the old system does not specify which action should expect an error, just that one action should raise it.  That is not particularly great in my opinion and this commit changes the new system so that as well as tying the error to a particular action (as it did before), actions declared after an action with an expected action will still be executed - the old system would instead exit early, the request in the old version of the converted test was never actually executed.