PR FIXUP - Fix subscription tests
Sorry about the late change, there was a bug in the subscription mapping from the old system that resulted in these silently passing.  It proved inconvienent to fix in the mapping as the old system did not specify the results for the triggering mutations which are required in the new system.  I decided it would be better to just convert the existing tests to the new system.