refactor: Generate gql types using col. desc (#1080)
* Remove commented out property

* Use correct func name in comment

* BREAKING - Remove schema.FieldIds

It is not used and creates a needless parsing complication (and effort for patches). Changes the Cids.

* Remove collection.Indexes

Is not used and complicates parsing logic.

* Generate collection descriptions from gql sdl

Removes the strangeness on restart where the collection descriptions are generated from the public gql types.  They are instead now created directly by parsing the GQL SDL. This allows for more interesting changes later, where our core internal types drive the gql query type generation, instead of the other way around.

* Remove gql/mutation test

Func signatures will change significantly quite soon, and this test is of too low value to rewrite (covered by integration tests).

* Remove unused Type type

* Generate gql types using collection descriptions

Simplifies the flow somewhat and allows our strongly typed model to be the primary driver.  This allows more work to leave the gql space, including the persisted [gql]schema in the datastore allowing for greater decoupling of Defra from gql. It also frees the generator code from having to worry about any internal database stuff such as relationship management, focusing on just gql stuff.

It also means that any Json patches applied to the persisted json schema (in /collection/version) will automatically be applied to the query language (gql) types. This /collection/version key will require some changes when we start to introduce local-only elements, however that is not a new problem to this PR.

It also decouples the gql types from the gql sdl, in theory allowing a non-gql sdl to generate gql types, or a gql sdl to generate types for a non-gql query language.

* Remove commented out code

* Move ast related funcs to gql/collection.go

They are out of place now as generate.go no longer touches the ast, and collection.go is the only place that calls them