Add P2P capacity to core test framework
Subscriptions, transactional queries and the change detector are currently out of scope, although I see no reason as to why they could not be made to work later.

Actually using these new actions has been split out to another commit in order to keep this on focused, it may however be handy to have a look at the two together to get an idea as to what it looks like to use it.