Move control of db instance to action handler
We need to be able to initialize database instances from within an action, but they should still be using the correct datastore type! We also require 1-many database instances for a given test (one per node).  This commit enables that, and creates a single 'normal' node if no nodes have been explicitly requested.

The explain utils starts to get in the way a little bit here, and will be properly separated shortly in another commit.