Implemented new client interface within DB package.
Ignore some of the names like `inner` and `outer`. Just put something on paper.

The priamry design here is to have all the actual CRUD related operations on the
client.DB interfaces to be implenented on a sub type. This sub type is embedded
within the db.db type.

It has all the necessary public functions. The original implementations for the
various operations are kept the exact same, but are (mostly) under the name
<operation>Txn which is a private function. These functions take a `txn` as a
parameter.

The `innerDB` type is responsible for the implicit/explicit transaction management.
Moreover, `db.WithTxn` returns an instance of the `innerDB` which implements the
`client.Store` interface. The returned `innerDB` has a manually set `txn` that
will be used for all internal operations.

Fixing any potentially broken tests to follow...