test: Migrate P2P/state tests to new framework (#1160)
* Remove replicator race condition

Flagged by go test -race

* Remove Defra-IM txn error fanciness

It does not work with errors.Is, which is kind of important and breaks P2Ps transaction rentry code

* Remove dbInfo.Rootstore func

Is an unnessecary complication that makes future changes harder

* Extract datastore strings to consts

databaseType will be used more in later commits.  It can stay internal and hidden behind the `Name()` function for the explain tests for now.

* Move control of db instance to action handler

We need to be able to initialize database instances from within an action, but they should still be using the correct datastore type! We also require 1-many database instances for a given test (one per node).  This commit enables that, and creates a single 'normal' node if no nodes have been explicitly requested.

The explain utils starts to get in the way a little bit here, and will be properly separated shortly in another commit.

* Remove databaseInfo from main test utils

databaseInfo and a handful of utils functions are now only used by the explain tests and represent an annoying code complication and oddity for the main test suite.  This commit removes what it can, and moves the rest into the explain test utils.

* Add P2P capacity to core test framework

Subscriptions, transactional queries and the change detector are currently out of scope, although I see no reason as to why they could not be made to work later.

Actually using these new actions has been split out to another commit in order to keep this on focused, it may however be handy to have a look at the two together to get an idea as to what it looks like to use it.

* Convert existing P2P tests to new system

Does not change the nature of the tests, although I do remember deleting a single test that acted as a sanity check for the old framework.

Does not make use of localised utils.go files in order to hide the schema - whilst they are consistent across a directory I am currently prefering the explicitness here given how complex these tests are.  Do let me know though.

* Remove old P2P/state test framework