Add support for update from sec. side
Note: This does not affect the collection.Update function, as that goes via `c.save`, and not `c.applyPatch`.  That behaviour will be changed with either mutation-create, or a different ticket for those calls. It might be nice to investigate unifying these at somepoint, but not now.

The two new functions isSecondaryIDField and patchPrimaryDoc will be used as-is when adding support for wrong-side save calls.  They are broken into two functions due to the differences in `c.save` and `c.applyPatch`.  It may be worth reviewing these two functions within the context of both commits.